Update
We're almost done with inputting all of the script text in Dialogic. Once complete, I think I'll focus on editing the text, adding the extra dialogue, and adding effects. Hopefully after that, we can try and debug the game and ask questions again in the server. That reminds me, I need to most likely redo the shop and that's going to suck.